+++ test/docs/activity_test.rb
@@ -12,10 +12,10 @@ class DocsIntermediateTest < Minitest::Spec
                                                  Intermediate::Out(:failure, :C)],
             Intermediate::TaskRef(:B)        => [Intermediate::Out(:success, :"End")],
             Intermediate::TaskRef(:C)        => [Intermediate::Out(:success, :B)],
-            Intermediate::TaskRef(:"End", stop_event: true) => [Intermediate::Out(:success, nil)] # :)
+            Intermediate::TaskRef(:"End", stop_event: true) => []
         },
-        [:"End"],   # end events
-        [:"Start"], # start
+        {:"End" => :success},   # termini
+        :"Start", # start
     )
     #:int end
 
@@ -47,7 +47,7 @@ class DocsIntermediateTest < Minitest::Spec
                                                               Activity::Output(Activity::Left, :failure)],  []),
         :B        => Implementation::Task(Upsert.method(:c), [Activity::Output(Activity::Right, :success)], []),
         :C        => Implementation::Task(Upsert.method(:c), [Activity::Output(Activity::Right, :success)], []),
-        :"End"    => Implementation::Task(_end, [Activity::Output(_end, :success)],                         []), # :)
+        :"End"    => Implementation::Task(_end, [],                         []), # :)
     }
     #:impl end